move litellm.integrations.vector_store_integrations
