+++ litellm/vector_stores/__init__.py
@@ -1,3 +1,4 @@
 from .main import asearch, search
+from .vector_store_registry import VectorStoreRegistry
 
-__all__ = ["search", "asearch"] 
+__all__ = ["search", "asearch", "VectorStoreRegistry"]
